requirements.txt
apache-airflow==2.9.1
requests
minio
python-dotenv